+++ src/mongodb/documents/meal-comment.document.ts
@@ -1,0 +1,8 @@
+import { Document } from 'mongoose';
+
+export interface MealCommentDocument extends Document {
+    readonly mealId: string;
+    readonly user: string;
+    readonly text: string;
+    readonly posted: number;
+}