+++ src/mongodb/schemas/meal-comment.schema.ts
@@ -1,0 +1,20 @@
+import * as mongoose from 'mongoose';
+
+export const MealCommentSchema = new mongoose.Schema({
+    mealId: {
+        type: String,
+        required: true
+    },
+    user: {
+        type: String,
+        required: true
+    },
+    text: {
+        type: String,
+        required: true
+    },
+    posted: {
+        type: Number,
+        required: true
+    }
+});